+++ src/app/inventory/item-move-service.ts
@@ -32,9 +32,10 @@ import {
 } from './dim-item-info';
 import reduxStore from '../store/store';
 import { count } from 'app/utils/util';
-import { itemInfosSelector } from './selectors';
+import { itemInfosSelector, itemHashTagsSelector } from './selectors';
 import { getStore, getItemAcrossStores, getCurrentStore, getVault } from './stores-helpers';
 import { touch } from './actions';
+import { ItemHashTag } from '@destinyitemmanager/dim-api-types';
 
 /**
  * You can reserve a number of each type of item in each store.
@@ -644,6 +645,7 @@ function ItemService(): ItemServiceType {
     }
 
     const itemInfos = itemInfosSelector(reduxStore.getState());
+    const itemHashTags = itemHashTagsSelector(reduxStore.getState());
 
     // A cached version of the space-left function
     const cachedSpaceLeft = _.memoize(
@@ -679,6 +681,7 @@ function ItemService(): ItemServiceType {
         store,
         targetStore,
         itemInfos,
+        itemHashTags,
         item
       ).find((candidate) => {
         const spaceLeft = cachedSpaceLeft(targetStore, candidate);
@@ -1021,6 +1024,9 @@ export function sortMoveAsideCandidatesForStore(
   fromStore: DimStore,
   targetStore: DimStore,
   itemInfos: ItemInfos,
+  itemHashTags: {
+    [itemHash: string]: ItemHashTag;
+  },
   /** The item we're trying to make space for. May be missing. */
   item?: DimItem
 ) {
@@ -1054,7 +1060,7 @@ export function sortMoveAsideCandidatesForStore(
       compareBy((i) => !fromStore.isVault && !i.canBeEquippedBy(fromStore)),
       // Tagged items sort by orders defined in dim-item-info
       compareBy((i) => {
-        const tag = getTag(i, itemInfos);
+        const tag = getTag(i, itemInfos, itemHashTags);
         return -(fromStore.isVault ? vaultDisplacePriority : characterDisplacePriority).indexOf(
           tag || 'none'
         );