+++ src/app/item-popup/ItemTagHotkeys.tsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import { itemTagList, TagValue, getTag } from '../inventory/dim-item-info';
 import { DimItem } from '../inventory/item-types';
-import { setItemTag } from '../inventory/actions';
+import { setItemTag, setItemHashTag } from '../inventory/actions';
 import { Hotkey } from '../hotkeys/hotkeys';
 import GlobalHotkeys from '../hotkeys/GlobalHotkeys';
 import { t } from 'app/i18next-t';
 import { connect } from 'react-redux';
-import { RootState } from 'app/store/reducers';
-import { itemInfosSelector } from 'app/inventory/selectors';
+import { RootState, ThunkDispatchProp } from 'app/store/reducers';
+import { itemInfosSelector, itemHashTagsSelector } from 'app/inventory/selectors';
+import { itemIsInstanced } from 'app/utils/item-utils';
 
 interface ProvidedProps {
   item: DimItem;
@@ -20,18 +21,13 @@ interface StoreProps {
 
 function mapStateToProps(state: RootState, props: ProvidedProps): StoreProps {
   return {
-    itemTag: getTag(props.item, itemInfosSelector(state)),
+    itemTag: getTag(props.item, itemInfosSelector(state), itemHashTagsSelector(state)),
   };
 }
 
-const mapDispatchToProps = {
-  setItemTag,
-};
-type DispatchProps = typeof mapDispatchToProps;
+type Props = ProvidedProps & StoreProps & ThunkDispatchProp;
 
-type Props = ProvidedProps & StoreProps & DispatchProps;
-
-function ItemTagHotkeys({ item, children, itemTag, setItemTag }: Props) {
+function ItemTagHotkeys({ item, children, itemTag, dispatch }: Props) {
   if (!item.taggable) {
     return <>{children}</>;
   }
@@ -44,7 +40,14 @@ function ItemTagHotkeys({ item, children, itemTag, setItemTag }: Props) {
         combo: tag.hotkey,
         description: t('Hotkey.MarkItemAs', { tag: tag.type }),
         callback: () =>
-          setItemTag({ itemId: item.id, tag: itemTag === tag.type ? undefined : tag.type }),
+          dispatch(
+            itemIsInstanced(item)
+              ? setItemTag({ itemId: item.id, tag: itemTag === tag.type ? undefined : tag.type })
+              : setItemHashTag({
+                  itemHash: item.hash,
+                  tag: itemTag === tag.type ? undefined : tag.type,
+                })
+          ),
       });
     }
   });
@@ -56,7 +59,4 @@ function ItemTagHotkeys({ item, children, itemTag, setItemTag }: Props) {
   );
 }
 
-export default connect<StoreProps, DispatchProps>(
-  mapStateToProps,
-  mapDispatchToProps
-)(ItemTagHotkeys);
+export default connect<StoreProps>(mapStateToProps)(ItemTagHotkeys);